requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
faiss-cpu>=1.9.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.7.0
tiktoken>=0.7.0
pandas>=2.0.0
openpyxl>=3.1.0
